+++ src/main/java/com/cs/os/cpuscheduler/FCFS.java
@@ -5,13 +5,13 @@ import java.util.Queue;
 
 class FCFSAlgorithm implements SchedulingAlgorithm {
     @Override
-    public Pair<Process, Float> schedule(Queue<Process> processQueue, CPU cpu) {
+    public Pair<Process, Integer> schedule(Queue<Process> processQueue, CPU cpu) {
         if (!processQueue.isEmpty()) {
 
             Process currentProcess = processQueue.poll();
 
             // Calculate waiting time
-            float executionTime = Math.max(0, currentProcess.getBurstTime() - currentProcess.getExecutionTime());
+            int executionTime = Math.max(0, currentProcess.getBurstTime() - currentProcess.getExecutionTime());
 
             // Execute the process on the CPU
             cpu.executeProcess(currentProcess, currentProcess.getBurstTime());
